Fix bug in selecting by name; Move errors; Fix minnehaha pkwy query;  update notebooks
