+++ examples/stpaul/project_cards/road.prop_change.widen.yml
@@ -1,20 +1,11 @@
 project: Widen Minnehaha Pkwy
+tags: ["baseline"]
 notes: This facility selection is done with specific model link IDs. It could also be done using from/to nodes and the name.
 changes:
 - roadway_property_change:
     facility:
       links:
-        model_link_id:
-          - 12436
-          - 46729
-          - 194599
-          - 266919
-          - 261737
-          - 362456
-          - 15165
-          - 362457
-          - 4382
-          - 12437
+        name: ["Minnehaha"]
     property_changes:
       lanes:
         existing: 1
