+++ network_wrangler/roadway/links/links.py
@@ -5,19 +5,12 @@ from typing import Optional
 import pandas as pd
 from pandera.typing import DataFrame
 
+from ...errors import LinkNotFoundError, MissingNodesError, NotLinksError
 from ...logger import WranglerLogger
 from ...models.roadway.tables import RoadLinksTable, RoadNodesTable, RoadShapesTable
 from ...utils.data import fk_in_pk
 from .filters import filter_links_not_in_ids, filter_links_to_ids
 from .validate import validate_links_have_nodes
-
-
-class NotLinksError(Exception):
-    """Raised when a dataframe is not a RoadLinksTable."""
-
-
-class LinkNotFoundError(Exception):
-    """Raised when a link is not found in the links table."""
 
 
 def node_ids_in_links(
